requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
bcrypt==4.1.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
distro==1.9.0
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.2
flatbuffers==23.5.26
fonttools==4.48.1
google-auth==2.28.0
google-auth-oauthlib==1.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.8.2
mediapipe==0.10.9
motor==3.3.2
mp==0.5.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.12.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
paramiko==3.4.0
pillow==10.2.0
pluggy==1.4.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.1
PyNaCl==1.5.0
pyparsing==3.1.1
pytest==8.1.1
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pyttsx3==2.90
pytube==15.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
resend==2.21.0
pytest-asyncio==0.23.8
aiosmtplib==3.0.1
